New version of runCode.sh with new "executor.c"
Make runCode.sh just compile the .c, .cpp code, check if compile error,
and then call executor. Use executor to moniter and execute the submit
program.
Executor can tell if the program execute successfully, or TLE, MLE, RE,
CE(for python only). TLE is for the timeout duration, not for the time
limit in problem. MLE is for the maximum memory usage, not for the
memory limit in problem. In this version, TLE is set 10 sec, MLE is set
1024 * 10 kB(10mB).
Method used currentlly to prevent the CPU usage get too high for
executor is just making it sleep for a while. Maybe it will be modified
to a better one in later version.

!!!!!!!!!!!!!!!!!WARNING!!!!!!!!!!!!!!!!!!!!!!
This version is not set limit to CPU usage for submit program. So if
you submit a program that will consume all CPU utilization, it's
highly possible cause the system crash.